food_fighters/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
pandas==1.5.1
django-cors-headers==1.1.0